Refactor: Remove remaining assessments module and update related code
