--- servers/assessment/assessments/remainingAssessments/main.go
@@ -1,16 +1,0 @@
-package remainingAssessments
-
-import (
-	"github.com/gin-gonic/gin"
-	"github.com/jackc/pgx/v5/pgxpool"
-	promptSDK "github.com/ls1intum/prompt-sdk"
-	db "github.com/ls1intum/prompt2/servers/assessment/db/sqlc"
-)
-
-func InitRemainingAssessmentsModule(routerGroup *gin.RouterGroup, queries db.Queries, conn *pgxpool.Pool) {
-	setupRemainingAssessmentsRouter(routerGroup, promptSDK.AuthenticationMiddleware)
-	RemainingAssessmentsServiceSingleton = &RemainingAssessmentsService{
-		queries: queries,
-		conn:    conn,
-	}
-}
