+++ servers/assessment/main.go
@@ -12,7 +12,6 @@ import (
 	promptSDK "github.com/ls1intum/prompt-sdk"
 	"github.com/ls1intum/prompt2/servers/assessment/assessments"
 	"github.com/ls1intum/prompt2/servers/assessment/assessments/assessmentCompletion"
-	"github.com/ls1intum/prompt2/servers/assessment/assessments/remainingAssessments"
 	"github.com/ls1intum/prompt2/servers/assessment/assessments/scoreLevel"
 	"github.com/ls1intum/prompt2/servers/assessment/categories"
 	"github.com/ls1intum/prompt2/servers/assessment/competencies"
@@ -93,7 +92,6 @@ func main() {
 	categories.InitCategoryModule(api, *query, conn)
 	assessments.InitAssessmentModule(api, *query, conn)
 	assessmentCompletion.InitAssessmentCompletionModule(api, *query, conn)
-	remainingAssessments.InitRemainingAssessmentsModule(api, *query, conn)
 	scoreLevel.InitScoreLevelModule(api, *query, conn)
 
 	serverAddress := promptSDK.GetEnv("SERVER_ADDRESS", "localhost:8084")
